css fais

+++ src/Controller/MediaController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Media;
+use App\Form\MediaType;
 use App\Repository\MediaRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -10,52 +11,42 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
+#[Route('/media')]
 class MediaController extends AbstractController
 {
-    /**
-     * @Route("/", name="app_home", methods={"GET"})
-     */
-    public function home(): Response
-    {
-        return $this->render('home.html.twig');
-    }
-
-    /**
-     * @Route("/media", name="app_media_index", methods={"GET"})
-     */
+    #[Route('/', name: 'app_media_index', methods: ['GET'])]
     public function index(MediaRepository $mediaRepository): Response
     {
-        $media = $mediaRepository->findAll();
+        // On récupère tous les médias depuis le repository
+        $medias = $mediaRepository->findAll();
+
+        // On rend la vue en passant la liste des médias
         return $this->render('media/index.html.twig', [
-            'media' => $media,
+            'media' => $medias,
         ]);
     }
 
-    /**
-     * @Route("/media/new", name="app_media_new", methods={"GET", "POST"})
-     */
+    #[Route('/new', name: 'app_media_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
-        if ($request->isMethod('POST')) {
-            $medium = new Media();
-            $medium->setTitle($request->request->get('title', ''));
-            $medium->setGroupName($request->request->get('group_name', ''));
-            $medium->setStyle($request->request->get('style', ''));
-            $medium->setYear((int)$request->request->get('year', 0));
-            $medium->setMedium($request->request->get('medium', ''));
-            
+        $medium = new Media();
+        $form = $this->createForm(MediaType::class, $medium);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->persist($medium);
             $entityManager->flush();
-            
-            return $this->redirectToRoute('app_media_index');
+
+            return $this->redirectToRoute('app_media_index', [], Response::HTTP_SEE_OTHER);
         }
-        
-        return $this->render('media/new.html.twig');
+
+        return $this->render('media/new.html.twig', [
+            'medium' => $medium,
+            'form' => $form,
+        ]);
     }
 
-    /**
-     * @Route("/media/{id}", name="app_media_show", methods={"GET"})
-     */
+    #[Route('/{id}', name: 'app_media_show', methods: ['GET'])]
     public function show(Media $medium): Response
     {
         return $this->render('media/show.html.twig', [
@@ -63,36 +54,34 @@ class MediaController extends AbstractController
         ]);
     }
 
-    /**
-     * @Route("/media/{id}/edit", name="app_media_edit", methods={"GET", "POST"})
-     */
+
+
+    #[Route('/{id}/edit', name: 'app_media_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Media $medium, EntityManagerInterface $entityManager): Response
     {
-        if ($request->isMethod('POST')) {
-            $medium->setTitle($request->request->get('title', ''));
-            $medium->setGroupName($request->request->get('group_name', ''));
-            $medium->setStyle($request->request->get('style', ''));
-            $medium->setYear((int)$request->request->get('year', 0));
-            $medium->setMedium($request->request->get('medium', ''));
-            
+        $form = $this->createForm(MediaType::class, $medium);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->flush();
-            
-            return $this->redirectToRoute('app_media_index');
+
+            return $this->redirectToRoute('app_media_index', [], Response::HTTP_SEE_OTHER);
         }
-        
+
         return $this->render('media/edit.html.twig', [
             'medium' => $medium,
+            'form' => $form,
         ]);
     }
 
-    /**
-     * @Route("/media/{id}/delete", name="app_media_delete", methods={"POST"})
-     */
-    public function delete(Media $medium, EntityManagerInterface $entityManager): Response
+    #[Route('/{id}', name: 'app_media_delete', methods: ['POST'])]
+    public function delete(Request $request, Media $medium, EntityManagerInterface $entityManager): Response
     {
-        $entityManager->remove($medium);
-        $entityManager->flush();
-        
-        return $this->redirectToRoute('app_media_index');
+        if ($this->isCsrfTokenValid('delete'.$medium->getId(), $request->request->getString('_token'))) {
+            $entityManager->remove($medium);
+            $entityManager->flush();
+        }
+
+        return $this->redirectToRoute('app_media_index', [], Response::HTTP_SEE_OTHER);
     }
 }
